refactor: consolidate weather card components and add backdrop control
- Move weather card variants into WeatherCard folder with index.tsx
- Integrate weather utilities (backgrounds, styles) into component
- Add flexible backdrop prop to GridCard for CSS variable control
- Remove separate utility files for better code locality
- Update all weather cards to disable backdrop with background images
