+++ src/components/WeatherCard/WeatherCardDefault.tsx
@@ -5,7 +5,7 @@ import { ErrorBoundary, SkeletonCard, ErrorDisplay } from '../ui'
 import { GridCardWithComponents as GridCard } from '../GridCard'
 import type { CardProps } from '../cardRegistry'
 import type { HassEntity, EntityAttributes } from '~/store/entityTypes'
-import { getWeatherBackground } from '~/utils/weatherBackgrounds'
+import { getWeatherBackground } from './index'
 
 interface WeatherAttributes extends EntityAttributes {
   temperature?: number
@@ -125,6 +125,7 @@ function WeatherCardDefaultContent(props: CardProps) {
         onDelete={onDelete}
         onConfigure={onConfigure}
         hasConfiguration={!!onConfigure}
+        backdrop={false}
       >
         <Flex direction="column" align="center" justify="center" gap="2">
           <GridCard.Icon>
@@ -167,6 +168,7 @@ function WeatherCardDefaultContent(props: CardProps) {
       onConfigure={onConfigure}
       hasConfiguration={!!onConfigure}
       title={isStale ? 'Weather data may be outdated' : undefined}
+      backdrop={!backgroundImage}
       style={{
         backgroundImage: backgroundImage ? `url(${backgroundImage})` : undefined,
         backgroundSize: 'cover',
