--- src/components/weather-variants/index.ts
@@ -1,4 +1,0 @@
-export { WeatherCardDefault } from './WeatherCardDefault'
-export { WeatherCardMinimal } from './WeatherCardMinimal'
-export { WeatherCardModern } from './WeatherCardModern'
-export { WeatherCardDetailed } from './WeatherCardDetailed'
